% Generated by roxygen2 (4.0.2): do not edit by hand
\name{motHessian}
\alias{motHessian}
\title{Hessian matrix of log-Likelihood for mot model}
\usage{
motHessian(param, x, y, tau)
}
\arguments{
\item{param}{parameter vector: (beta_0, beta_1, ... , beta_m, sigma).}

\item{x}{design matrix of the model.}

\item{y}{observation vector.}

\item{tau}{threshold vector from tau_1 to tau_K.}
}
\value{
hessian matrix, summarized over all observations.
}
\description{
Hessian matrix of log-Likelihood for right censored Multiple Ordinal Tobit (MOT) model.
}
\author{
Marvin Wright
}
\seealso{
\link[lmmot]{lmmot}
}

