% Generated by roxygen2 (4.0.2): do not edit by hand
\name{summary.lmmot}
\alias{summary.lmmot}
\title{Summary if lmmot object}
\usage{
\method{summary}{lmmot}(object, digits = max(3, getOption("digits") - 3), ...)
}
\arguments{
\item{object}{lmmot object to print.}

\item{digits}{number of decimal digits to print.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print details about lmmot object.
}
\author{
Marvin Wright
}
\seealso{
\link[stats]{lm} \link[lmmot]{lmmot}
}

